Tested: works with updates and generated event qr code
