+++ app/src/main/res/layout/fragment_qr_code_scanner.xml
@@ -7,7 +7,6 @@
     android:layout_height="match_parent"
     android:background="@color/backgroundColor">
 
-<<<<<<< HEAD
     <!-- Barcode scanner view -->
     <com.journeyapps.barcodescanner.DecoratedBarcodeView
         android:id="@+id/barcode_scanner"
@@ -15,21 +14,6 @@
         android:layout_height="match_parent">
 
     </com.journeyapps.barcodescanner.DecoratedBarcodeView>
-=======
-    <!--    <com.journeyapps.barcodescanner.BarcodeView-->
-    <!--        android:id="@+id/barcode_scanner"-->
-    <!--        android:layout_width="wrap_content"-->
-    <!--        android:layout_height="wrap_content"-->
-    <!--        app:layout_constraintBottom_toBottomOf="parent"-->
-    <!--        app:layout_constraintEnd_toEndOf="parent"-->
-    <!--        app:layout_constraintHorizontal_bias="0.0"-->
-    <!--        app:layout_constraintStart_toStartOf="parent"-->
-    <!--        app:layout_constraintTop_toTopOf="parent"-->
-    <!--        app:layout_constraintVertical_bias="0.0">-->
-
->>>>>>> 73342bc90c98c2cf5d33e7f74ba48a64c632a94d
-
-    <!--    </com.journeyapps.barcodescanner.BarcodeView>-->
 
     <ImageView
         android:id="@+id/back_arrow"
